requirements.txt
psycopg2-binary==2.9.9
pglast==6.2
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
boto3==1.34.0
pydantic==2.5.3
